python/facerec/requirements.txt
dlib==19.10.0
opencv_python==3.4.0.12
py_postgresql==1.2.1
face_recognition==1.2.2
numpy==1.14.2
typing==3.6.4
PyYAML==3.12
